+++ tests/testthat/test-os-helper-functions.R
@@ -1,0 +1,4 @@
+test_that("OS tests", {
+  libs = get_installed_libs()
+  expect_gt(length(libs), 1000)
+})